#129 [BE] 랭킹 도메인 분리 및 누락 테스트 코드 작성 (#130)
* feat: Ranking 앱 생성

* chore: Ranking 앱 전역 등록

* refactor: Account 앱의 Ranking 로직 분리

* refactor: Home Ranking API 이름, 에러처리 방식 변경

* refactor: create_user 테스트 코드에서 user_score, user_solved 모델을 자동 생성하도록 변경

* test: Ranking 앱 테스트 코드 작성
