+++ backend/account/serializers.py
@@ -1,5 +1,4 @@
 from django import forms
-from django.core.validators import RegexValidator
 
 from utils.api import serializers, UsernameSerializer
 
@@ -21,14 +20,6 @@ class HomeStatistics(serializers.Serializer):
     total_problem_length = serializers.IntegerField()
     accepted_problem_length = serializers.IntegerField()
     ended_contest_length = serializers.IntegerField()
-
-
-class RankingSerializer(serializers.ModelSerializer):
-    class Meta:
-        user = serializers.ReadOnlyField(source="User.id")
-
-        model = UserScore
-        fields = ["score", "fluctuation", "user", "basis"]
 
 
 class UserLoginSerializer(serializers.Serializer):
@@ -125,73 +116,6 @@ class UserProfileSerializer(serializers.ModelSerializer):
         return obj.real_name if self.show_real_name else None
 
 
-class UserRankListSerializer(serializers.ModelSerializer):
-    rank = serializers.SerializerMethodField()
-    avatar = serializers.SerializerMethodField()
-    username = serializers.CharField()  # 수정
-    mood = serializers.CharField(source='userprofile.mood')  # 수정
-    score = serializers.IntegerField(source='userscore.total_score')  # 수정
-    major = serializers.CharField(source='userprofile.major')  # 수정
-    tier = serializers.CharField(source='userscore.tier')  # 수정
-    solved = serializers.SerializerMethodField()
-    growth = serializers.SerializerMethodField()
-
-    class Meta:
-        model = User
-        fields = ['rank', 'avatar', 'username', 'mood', 'score', 'major', 'tier', 'solved', 'growth']
-
-    def get_rank(self, obj):
-        return self.context['rank']
-
-    def get_avatar(self, obj):
-        return obj.userprofile.avatar
-
-    def get_growth(self, obj):
-        return obj.userscore.fluctuation
-
-    def get_solved(self, obj):  # 추가
-        return obj.usersolved.total_solved
-
-
-class SurgeUserSerializer(serializers.ModelSerializer):
-    rank = serializers.SerializerMethodField()
-    avatar = serializers.SerializerMethodField()
-    username = serializers.CharField()
-    mood = serializers.CharField(source='userprofile.mood')
-    score = serializers.IntegerField(source='userscore.total_score')
-    major = serializers.CharField(source='userprofile.major')
-    tier = serializers.CharField(source='userscore.tier')
-    solved = serializers.SerializerMethodField()
-    growth = serializers.SerializerMethodField()
-
-    class Meta:
-        model = User
-        fields = ['rank', 'avatar', 'username', 'mood', 'score', 'major', 'tier', 'solved', 'growth']
-
-    def get_rank(self, obj):
-        return self.context['rank']
-
-    def get_avatar(self, obj):
-        return obj.userprofile.avatar
-
-    def get_solved(self, obj):
-        return obj.usersolved.total_solved
-
-    def get_growth(self, obj):
-        return obj.userscore.fluctuation
-
-
-class MajorRankListSerializer(serializers.ModelSerializer):
-    rank = serializers.IntegerField()
-    major = serializers.CharField()
-    score = serializers.IntegerField()
-    people = serializers.IntegerField()
-
-    class Meta:
-        model = Department
-        fields = ['rank', 'major', 'score', 'people']
-
-
 class DashboardUserInfoSerializer(serializers.ModelSerializer):
     class Meta:
         model = User
@@ -284,21 +208,6 @@ class DashboardDifficultyInfoSerializer(serializers.ModelSerializer):
             }
 
         return difficulty_info
-
-
-class HomeRankingSerializer(serializers.ModelSerializer):
-    avatar = serializers.SerializerMethodField()
-    username = serializers.SerializerMethodField()
-
-    class Meta:
-        model = UserScore
-        fields = ['avatar', 'tier', 'username', 'total_score', 'fluctuation']
-
-    def get_avatar(self, obj):
-        return obj.user.userprofile.avatar
-
-    def get_username(self, obj):
-        return obj.user.username
 
 
 class EditUserSerializer(serializers.Serializer):
