+++ backend/account/urls/oj.py
@@ -1,14 +1,14 @@
 from django.conf.urls import url
 
-from ..views.oj import (ApplyResetPasswordAPI, ResetPasswordAPI, GetRankingAPI,
+from ..views.oj import (ApplyResetPasswordAPI, ResetPasswordAPI,
                         UserChangePasswordAPI, UserRegisterAPI,
                         UserLoginAPI, UserLogoutAPI, UsernameOrEmailCheck,
                         AvatarUploadAPI, TwoFactorAuthAPI, UserProfileAPI,
                         UserProfileDashBoardAPI,
-                        UserRankAPI, CheckTFARequiredAPI, SessionManagementAPI,
+                        CheckTFARequiredAPI, SessionManagementAPI,
                         ProfileProblemDisplayIDRefreshAPI, OpenAPIAppkeyAPI, SSOAPI,
                         ApplyUserEmailValidCheckAPI, UserEmailValidCheckAPI,
-                        GetCollegeListAPI, GetDepartmentListAPI, HomeRankingAPI, SurgeUserRankAPI, MajorRankAPI,
+                        GetCollegeListAPI, GetDepartmentListAPI,
                         GetHomeStatisticsAPI, NicknameValidCheckAPI)
 
 from utils.captcha.views import CaptchaAPIView
@@ -17,8 +17,6 @@ urlpatterns = [
     url(r"^college_list/?$", GetCollegeListAPI.as_view(), name="college_list"),
     url(r"^department_list/?$", GetDepartmentListAPI.as_view(), name="department_list"),
     url(r"^home_statistics/?$", GetHomeStatisticsAPI.as_view(), name="home_statistics"),
-    url(r"^ranking/?$", GetRankingAPI.as_view(), name="ranking"),
-    url(r"^home_ranking/?$", HomeRankingAPI.as_view(), name="home_ranking"),
     url(r"^login/?$", UserLoginAPI.as_view(), name="user_login_api"),
     url(r"^logout/?$", UserLogoutAPI.as_view(), name="user_logout_api"),
     url(r"^apply_user_email_valid_check/?$", ApplyUserEmailValidCheckAPI.as_view(),
@@ -37,9 +35,6 @@ urlpatterns = [
     url(r"^upload_avatar/?$", AvatarUploadAPI.as_view(), name="avatar_upload_api"),
     url(r"^tfa_required/?$", CheckTFARequiredAPI.as_view(), name="tfa_required_check"),
     url(r"^two_factor_auth/?$", TwoFactorAuthAPI.as_view(), name="two_factor_auth_api"),
-    url(r"^user_rank/?$", UserRankAPI.as_view(), name="user_rank_api"),
-    url(r"^surge_user_rank/?$", SurgeUserRankAPI.as_view(), name="surge_user_rank_api"),
-    url(r"^major_rank/?$", MajorRankAPI.as_view(), name="major_rank_api"),
     url(r"^sessions/?$", SessionManagementAPI.as_view(), name="session_management_api"),
     url(r"^open_api_appkey/?$", OpenAPIAppkeyAPI.as_view(), name="open_api_appkey_api"),
     url(r"^sso?$", SSOAPI.as_view(), name="sso_api")
