+++ backend/account/views/oj.py
@@ -31,11 +31,11 @@ from ..serializers import (ApplyResetPasswordSerializer, ResetPasswordSerializer
                            UserChangePasswordSerializer, UserLoginSerializer,
                            UserRegisterSerializer, UsernameOrEmailCheckSerializer,
                            RankInfoSerializer, UserChangeEmailSerializer, SSOSerializer,
-                           CollegeListSerializer, DepartmentSerializer, RankingSerializer,
+                           CollegeListSerializer, DepartmentSerializer,
                            DashboardSubmissionSerializer,
                            DashboardDepartmentSerializer, DashboardCollegeSerializer, DashboardRankSerializer,
-                           HomeRankingSerializer, DashboardUserInfoSerializer, DashboardFieldInfoSerializer,
-                           DashboardDifficultyInfoSerializer, UserRankListSerializer, SurgeUserSerializer,
+                           DashboardUserInfoSerializer, DashboardFieldInfoSerializer,
+                           DashboardDifficultyInfoSerializer,
                            HomeStatistics)
 from ..serializers import (TwoFactorAuthCodeSerializer, UserProfileSerializer,
                            EditUserProfileSerializer, ImageUploadForm)
@@ -86,18 +86,6 @@ class GetHomeStatisticsAPI(APIView):
         }
 
         return self.success(HomeStatistics(home_statistics).data)
-
-class GetRankingAPI(APIView):
-    def get(self, request):
-        try:
-            limit = int(request.GET.get("limit", None))
-            if limit:
-                ranking = UserScore.objects.all().order_by('-score')[:limit]
-            else:
-                ranking = UserScore.objects.all().order_by('-score')
-            return self.success(RankingSerializer(ranking, many=True).data)
-        except UserScore.DoesNotExist:
-            return HttpResponseNotFound("no ranking table")
 
 
 class UserProfileAPI(APIView):
@@ -181,17 +169,6 @@ class UserProfileDashBoardAPI(APIView):
             'difficultyInfo': difficultyInfo
         }
         return self.success(response_data)
-
-
-class HomeRankingAPI(APIView):
-    def get(self, request):
-        # rank, avatar, tier, total_score, fluctuation
-        limit = request.GET.get('limit', 100)
-        try:
-            ranking = UserScore.objects.all().order_by('-total_score')[:limit]
-        except UserScore.DoesNotExist:
-            return HttpResponseNotFound('no user score table')
-        return self.success(HomeRankingSerializer(ranking, many=True).data)
 
 
 class AvatarUploadAPI(APIView):
@@ -549,98 +526,6 @@ class SessionManagementAPI(APIView):
             return self.success("Succeeded")
         else:
             return self.error("Invalid session_key")
-
-
-class UserRankAPI(APIView):
-    # def get(self, request):
-    #     rule_type = request.GET.get("rule")
-    #     if rule_type not in ContestRuleType.choices():
-    #         rule_type = ContestRuleType.ACM
-    #     profiles = UserProfile.objects.filter(user__admin_type=AdminType.REGULAR_USER, user__is_disabled=False) \
-    #         .select_related("user")
-    #     if rule_type == ContestRuleType.ACM:
-    #         profiles = profiles.filter(submission_number__gt=0).order_by("-accepted_number", "submission_number")
-    #     else:
-    #         profiles = profiles.filter(total_score__gt=0).order_by("-total_score")
-    #     return self.success(self.paginate_data(request, profiles, RankInfoSerializer))
-    def get(self, request):
-
-        offset = int(request.GET.get("offset", 0))
-        limit = int(request.GET.get("limit", 10))
-
-        users = User.objects.prefetch_related('userprofile', 'userscore', 'usersolved') \
-                    .order_by('-userscore__total_score')[offset:offset + limit]
-
-        total_users = User.objects.count()
-
-        results = []
-        for rank, user in enumerate(users, start=offset + 1):
-            serializer = UserRankListSerializer(user, context={'rank': rank})
-            results.append(serializer.data)
-
-        data = {
-            'total': total_users,
-            'results': results
-        }
-
-        return self.success(data)
-
-class SurgeUserRankAPI(APIView):
-    def get(self, request):
-        offset = int(request.GET.get("offset", 0))
-        limit = int(request.GET.get("limit", 10))
-
-        users = User.objects.prefetch_related('userprofile', 'userscore', 'usersolved') \
-                            .order_by('-userscore__fluctuation')[offset:offset+limit]
-
-        total_users = User.objects.count()
-
-        results = []
-        for rank, user in enumerate(users, start=offset+1):
-            serializer = SurgeUserSerializer(user, context={'rank': rank})
-            results.append(serializer.data)
-
-        data = {
-            'total': total_users,
-            'results': results
-        }
-
-        return self.success(data)
-
-class MajorRankAPI(APIView):
-    def get(self, request):
-        limit = int(request.GET.get("limit", 7))
-
-        major_ranks = Department.objects.annotate(
-            score=Sum('userprofile__user__userscore__total_score'),
-            people=Count('userprofile__user')
-        ).filter(score__isnull=False).order_by('-score')[:limit]
-
-        total_majors = major_ranks.count()
-
-        results = []
-        for rank, major in enumerate(major_ranks, start=1):
-            people_data = major.userprofile_set.select_related('user', 'user__userscore').annotate(
-                avatar_url=F('user__userprofile__avatar'),
-                username=F('user__username'),
-                score=F('user__userscore__total_score'),
-                tier=F('user__userscore__tier')
-            ).order_by('-user__userscore__total_score').values('avatar_url', 'username', 'mood', 'score', 'tier')
-
-            data = {
-                'rank': rank,
-                'major': major.department_name,
-                'score': major.score,
-                'people': list(people_data)
-            }
-            results.append(data)
-
-        data = {
-            'total': total_majors,
-            'results': results
-        }
-
-        return self.success(data)
 
 class ProfileProblemDisplayIDRefreshAPI(APIView):
     @login_required
