+++ backend/oj/urls.py
@@ -17,4 +17,5 @@ urlpatterns = [
     url(r"^api/", include("banner.urls.oj")),
     url(r"^api/admin/", include("banner.urls.admin")),
     url(r"^api/", include("profile.urls.oj")),
+    url(r"^api/", include("ranking.urls.oj")),
 ]
